todos los tests puestos en tests.py, falta readaptarlos en las urls correspondientes.
