--- WebProjecte/requirements_test.txt
@@ -1,5 +1,0 @@
-Django==5.1.6
-selenium==4.15.2
-webdriver-manager==4.0.1
-pytest==7.4.3
-pytest-django==4.7.0
